% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_cubes.R
\name{add_cubes_to_scene}
\alias{add_cubes_to_scene}
\title{Add cubes to an existing scene}
\usage{
add_cubes_to_scene(scene, cubes, material = lambertian(color = "ivory"))
}
\arguments{
\item{scene}{scene}

\item{cubes}{cubes calculated with \code{mesh_to_cubes}}

\item{material}{1 material or a tibble of materials (1 row for each cube) from rayrender. lambertian, dielectric}
}
\value{
rayrender scene with cubes
}
\description{
Add cubes to an existing scene
}
\examples{
\dontrun{
library(Rvcg)
simple_mesh <- Rvcg::vcgIcosahedron()
cubes <- mesh_to_cubes(simple_mesh, voxelSize = 0.5, scene_dim = c(80, 475))

library(rayrender)
scene <- generate_cornell(lightintensity = 10)
# _add cubes on scene
scene <- mesh2ray::add_cubes_to_scene(scene, cubes = cubes,
                                     material = lambertian(color = "ivory"))
render_scene(scene, lookfrom = c(278, 278, -800) ,
  lookat = c(278, 278, 0), fov = 40, ambient_light = FALSE,
  samples = 500, parallel = TRUE, clamp_value = 5)
}

}
