% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_raster.R
\name{mesh_to_raster}
\alias{mesh_to_raster}
\title{Mesh3d to raster}
\usage{
mesh_to_raster(x, res = 150)
}
\arguments{
\item{x}{mesh3d object}

\item{res}{resolution of the resulting raster. 1 or 2 values for nrows, ncols.}
}
\value{
A raster
}
\description{
Mesh3d to raster
}
\examples{
library(Rvcg)
data(humface)
r <- mesh_to_raster(humface)
}
