% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_to_ray.R
\name{stack_to_ray}
\alias{stack_to_ray}
\title{Image and elevation to Rayshader}
\usage{
stack_to_ray(elevation, image = NULL)
}
\arguments{
\item{elevation}{Raster for elevation or path to elevation raster with same number of rows/cols}

\item{image}{RasterStack or path to png image. If not same dimension, image is resampled}
}
\value{
A list with 'elevation' for the elevation matrix and 'overlay' for the rgb overlay from the 'image'
}
\description{
Image and elevation to Rayshader
}
